[#16] Adding Info Boxes
What: Add info boxes for
- The teleporter
- The encoder
- The decoder
- The fire hazard
- Bell pairs

Source for the Dialog System
used for the info boxes:
https://godotengine.org/asset-library/asset/628

PS2:
- Removed decoder info box code.
This will be added when it's description
is finalized.
- Finished info boxes for the
remaining components.
- Removed the SPACEBAR key from
shooting qubits so the Otter won't shoot
when going through info box nodes

- TODO: Need to handle case where
multiple info boxes are available at
the same time

- TODO: Since the SPACEBAR key might
be most intuitive for shooting, re-introduce
it for shooting and handle the case when the
info box is open