[#17, #16] Add Sound Effects + Decoder Info Box + Nits and Bug Fixes
What:
- Added Otter die and
Boulder Explosion effects (and
connected the Right Boulder to
explode once the decoder is used)
- Added the Decoder Info Box
- Temporarily fixed bug where
the collectable blue bit info icon
appears over the fire trap info icon
by only showing the fire trap info icon
since there are two collectable bits so
the other can offer info
- Edited Otter sprite sheet so smoke
from shooting blows north east and west
and since the side and front-diagonal
animations are the same, made the smoke
effect close to the face to look more generic

=================
PS history of #16 Adding Info Boxes
=================
What: Add info boxes for
- The teleporter
- The encoder
- The decoder
- The fire hazard
- Bell pairs

Source for the Dialog System
used for the info boxes:
https://godotengine.org/asset-library/asset/628

PS2:
- Removed decoder info box code.
This will be added when it's description
is finalized.
- Finished info boxes for the
remaining components.
- Removed the SPACEBAR key from
shooting qubits so the Otter won't shoot
when going through info box nodes

- TODO: Need to handle case where
multiple info boxes are available at
the same time

- TODO: Since the SPACEBAR key might
be most intuitive for shooting, re-introduce
it for shooting and handle the case when the
info box is open